requirements.txt
openai==0.27.0
notion-client==2.0.0
flask==2.1.2
python-dotenv==1.0.0